server/requirements.txt
grpcio==1.62.2
grpcio-tools==1.62.2
numpy==1.26.4
protobuf==4.25.0
# opencv-python==4.9.0.80
#######################################
# opencv-python==4.8.0.74
